This works on all compilers. Timing info below:

```
Before
gfortran : 1543.38 sec
ifort    :  494.28 sec
nagfor   :  453.12 sec

After    :    x /= x   | ieee_is_nan
gfortran :  453.61 sec |  599.47 sec
ifort    :  482.28 sec |  478.88 sec
nagfor   :  438.23 sec |  446.21 sec
```

Before here is current master.
After, 1st column uses x /= x to check for NaN for new isnan_mb function, 2nd column uses ieee_is_nan.

For ifort and nagfor there is minimal difference. For gfortran the difference is in the aperture check which uses NaN checks a lot.

The PR manually inlines NaN checks internally in mathlib_bouncer, which is why the test time drops from 1543 sec to 599 sec for gfortran.